+++ Sources/WordPressSharedObjC/Utility/NSString+Helpers.m
@@ -1,12 +1,7 @@
 #import "NSString+Helpers.h"
 #import <CommonCrypto/CommonDigest.h>
 #import "NSString+XMLExtensions.h"
-
-#if SWIFT_PACKAGE
-#import "../Private/WPSharedLoggingPrivate.h"
-#else
-#import "WPSharedLoggingPrivate.h"
-#endif
+#import "WPSharedLogging.h"
 
 static NSString *const Ellipsis =  @"\u2026";
 
@@ -20,7 +15,7 @@ static NSString *const Ellipsis =  @"\u2026";
 + (NSString *)emojiFromCoreEmojiImageTag:(NSString *)tag
 {
     if ([tag rangeOfString:@"<img"].location == NSNotFound || [tag rangeOfString:@"/images/core/emoji/"].location == NSNotFound) {
-        DDLogError(@"Tried to extract emoji from a string that was not a core emoji image tag.");
+        WPSharedLogError(@"Tried to extract emoji from a string that was not a core emoji image tag.");
         return nil;
     }
 
@@ -100,7 +95,7 @@ static NSString *const Ellipsis =  @"\u2026";
         NSString *pattern = @"\\[[^\\]]+\\]";
         regex = [[NSRegularExpression alloc] initWithPattern:pattern options:NSRegularExpressionCaseInsensitive error:&error];
         if (error) {
-            DDLogError(@"Error parsing regex: %@", error);
+            WPSharedLogError(@"Error parsing regex: %@", error);
         }
     });
     NSRange range = NSMakeRange(0, [string length]);